+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Changed
+- update check-graylog2-alive.rb to accept an `--apipath` argument to specify the path of the transport api
+- update -alive lifecycle check because there are now multiple valid lifecyle states
+- update sensu-plugin dep to '~> 1.2'
+- add some tests, this pulled in webmock to mock restclient calls
+- update readme
 
 ## [0.1.0] - 2016-01-29
 ### Added